+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.1.4] - 2025-03-31
+### BREAKING CHANGE
+- Parse data based on device instead of model
+    - First supported device type is "controller"
+    - Allows for faster expansion to new supported devices
+
+### Improvements
+- Migrate to use pytest instead of unittest
 ## [0.1.3] - 2025-03-19
 ### Fixed
 - Fixed changelog URL in pyproject.toml